Changes the print function to one more generalist
Uses a size_t argument to be able to use one function anywhere. Changes
the Makefile to actually delete main.o. Changes functions prototypes to
use standard types.

+++ src/allocating.h
@@ -12,12 +12,14 @@
 #ifndef __ALLOCATING__
 #define __ALLOCATING__
 
+#include <stdint.h>
+
 /*
  * Defines to be used in alloc_image function.
  */
-#define CHAR 1
-#define SHORT 2
-#define INT 4
+#define CHAR ((size_t) 1)
+#define SHORT ((size_t) 2)
+#define INT ((size_t) 4)
 
 /*
  * Allocate a two-dimensional array filled with zeroes.
@@ -26,7 +28,7 @@
  * The function returns the address of the array on success and
  * NULL on error.
  */
-void **alloc_image(int lines, int columns, size_t bytes_per_element);
+void **alloc_image(uint32_t lines, uint32_t columns, size_t bytes_per_element);
 /*
  * Free the array containing the image.
  */
